+++ ui/src/pages/index.ts
@@ -1,6 +1,7 @@
 export * from "./Dashboard";
 export * from "./Txt2Img";
 export * from "./Img2Img";
+export * from "./Flux2Edit";
 export * from "./Inpaint";
 export * from "./ControlNet";
 export * from "./Kontext";